"RLC Glass"
goal is to add the ability for the user to select a custom background.

in order to reach this goal, the colors for RLC have been redone with a focus on using transparent backgrounds as much as possible, and letting the background determine the visuals.

for dark/non dark mode this means that background is controlled by the rlc wrapper, and text is made white for dark mode and black for non dark mode.

borders have been adjusted overall to be thinner

custom background image option prompts user when enabled to provide a url(while also providing a sample for the user)
this url is then injected via a style tag. this style tag is used instead of addstyle so we can change its value on the fly.

while adding the custom option, i discovered that the options toggle twice whenever activated by clicking on the label rather than the checkbox.
turns out that when the label is clicked, the click event is triggered twice. i have made a hacky adjustment to prevent the event from doing anything if the label was clicked, forcing user to click checkbox to interact with settings.

non dark mode is a bit meh.